*: a few uint32_t warnings
svn:sync-xref-src-repo-rev: r85565

+++ src/VBox/Main/glue/string.cpp
@@ -1,4 +1,4 @@
-/* $Id: string.cpp 44528 2013-02-04 14:27:54Z noreply@oracle.com $ */
+/* $Id: string.cpp 45927 2013-05-07 08:05:43Z noreply@oracle.com $ */
 /** @file
  * MS COM / XPCOM Abstraction Layer - UTF-8 and UTF-16 string classes.
  */
@@ -57,7 +57,7 @@ void Bstr::copyFromN(const char *a_pszSrc, size_t a_cchMax)
         throw std::bad_alloc();
     }
 
-    m_bstr = ::SysAllocStringByteLen(NULL, cwc * sizeof(OLECHAR));
+    m_bstr = ::SysAllocStringByteLen(NULL, (unsigned)(cwc * sizeof(OLECHAR)));
     if (RT_UNLIKELY(!m_bstr))
         throw std::bad_alloc();
